Upload of code for session of 13.05.2022
Upload of the code that was produced in the session on "Data preprocessing for Machine Learning" on 13.05.2022. The code is a "clean" version of what was shown during the session and also there are 2 scripts for both datasets that will be used furtheron. Note that there are no major differences between the 2 scripts besides that the preprocessing for the regression problem involves less steps in this case.